[SCD-70] refactoring, change directory structure
