+++ src/Controller/Diary/DeleteDiary/Controller.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Controller\Diary\DeleteDiary;
+
+use App\Entity\Diary;
+use App\Service\Diary\DiaryService;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+
+final class Controller extends AbstractController
+{
+    private DiaryService $diaryService;
+
+    public function __construct(DiaryService $diaryService)
+    {
+        $this->diaryService = $diaryService;
+    }
+
+    /**
+     * @Route("/api/diary/{noted_at}", name="delete_diary", methods={"DELETE"})
+     */
+    public function deleteDiary(?Diary $diary): Response
+    {
+        if (null === $diary) {
+            return new Response('', Response::HTTP_NOT_FOUND);
+        }
+        $this->diaryService->delete($diary);
+
+        return new Response('', Response::HTTP_NO_CONTENT);
+    }
+}
