+++ src/Controller/Diary/DiaryDTO.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\Diary;
 
 use Symfony\Component\Validator\Constraints as Assert;
