+++ src/Service/Diary/DiaryService.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Service;
+namespace App\Service\Diary;
 
-use App\Controller\DiaryDTO;
+use App\Controller\Diary\DiaryDTO;
 use App\Entity\Diary;
 use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
